refactor(lotto_service): 로또 서비스 리팩토링
작업 내용
- `LottoService`에서 처리하던 당첨 등수 계산 로직을 별도 도메인에서 처리하도록 수정
- 이율 계산 전용 도메인 생성 (`Scoreboard`)
